services/ingest/requirements.txt
requests
psycopg2-binary
prometheus-client
python-dateutil
pandas
openpyxl
xlrd